# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Psych Manager Origin API",
        "contact": {
            "email": "info@ps.ru"
        },
        "version": "1.0 beta"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "PMO server"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Категории"
                ],
                "summary": "Index page",
                "operationId": "categories-index",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Список пуст"
                    }
                },
                "security": [
                    {
                        "X-ACCESS-TOKEN": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Категории"
                ],
                "summary": "Store",
                "operationId": "categories-store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Категория добавлена"
                    },
                    "500": {
                        "description": "Ошибка создания"
                    }
                },
                "security": [
                    {
                        "X-ACCESS-TOKEN": []
                    }
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Категории"
                ],
                "summary": "Show page",
                "operationId": "categories-show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Название категории"
                    },
                    "404": {
                        "description": "Категория не найдена"
                    }
                },
                "security": [
                    {
                        "X-ACCESS-TOKEN": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Категории"
                ],
                "summary": "Update",
                "operationId": "categories-update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/CategoryRequest"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Категория обновлена"
                    },
                    "404": {
                        "description": "Категория не найдена"
                    }
                },
                "security": [
                    {
                        "X-ACCESS-TOKEN": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Категории"
                ],
                "summary": "Delete",
                "operationId": "categories-delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно"
                    },
                    "404": {
                        "description": "Категория не найдена"
                    },
                    "500": {
                        "description": "Ошибка удаления"
                    }
                },
                "security": [
                    {
                        "X-ACCESS-TOKEN": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CategoryRequest": {
                "properties": {
                    "name": {
                        "description": "Category name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "X-API-KEY": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}